% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_quantiles_maxk.R
\name{estimate_quantiles_maxk}
\alias{estimate_quantiles_maxk}
\title{Estimate Quantiles with Maxk}
\usage{
estimate_quantiles_maxk(samp = NULL, maxk_line = NULL, probs_interest = NULL)
}
\arguments{
\item{samp}{sample}

\item{maxk_line}{maxk line obtained for the probabilities of interest}

\item{probs_interest}{Probabilities of interest to estimate}
}
\value{
Returns the estimation of the quantiles using the maxk line
}
\description{
\code{estimate_quantiles_maxk} use the maxk line obtained to estimate quantiles with MIK
}
\examples{
linear_adjust(min_maxk = c(10, 15, 20),
              probs = c(1-1e-1, 1-1e-2, 1-1e-3),
              probs_interest = c(1-1e-6, 1-1e-7, 1-1e-8))
}
\keyword{RESTK}
