% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_min_maxk.R
\name{get_min_maxk}
\alias{get_min_maxk}
\title{Get the minimum maxk}
\usage{
get_min_maxk(samp_tightness = NULL, k_seq = NULL)
}
\arguments{
\item{samp_tightness}{tightness from a given sample and maxk}

\item{k_seq}{sequence of maxk to evaluate}
}
\value{
Returns the minimum maxk
}
\description{
\code{get_min_maxk} get the minimum maxk from a set of maxks and tightness
}
\examples{
get_min_maxk(samp_tightness = c(1.5, 1.2, 0.98),
             k_seq = c(20, 30 , 40))
}
\keyword{RESTK}
