% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_adjust.R
\name{linear_adjust}
\alias{linear_adjust}
\title{Linear adjust}
\usage{
linear_adjust(min_maxk = NULL, probs = NULL, probs_interest)
}
\arguments{
\item{min_maxk}{minimum maxk found for each probability of interest}

\item{probs}{Probabilities where maxk was evaluated}

\item{probs_interest}{Probabilities of interest to estimate}
}
\value{
Returns the maxk line for the probabilities of interest
}
\description{
\code{linear_adjust} function used to project the max_k line into the probabilities of interest
}
\examples{
linear_adjust(min_maxk = c(10, 15, 20),
              probs = c(1-1e-1, 1-1e-2, 1-1e-3),
              probs_interest = c(1-1e-6, 1-1e-7, 1-1e-8))
}
\keyword{RESTK}
